backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pydantic[email-validator]==2.10.4
email-validator==2.2.0
python-multipart==0.0.20
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
google-auth==2.37.0
requests==2.32.3
redis[hiredis]==5.2.1
websockets==14.1
